Move some common code from renderer/App.tsx to app/App.tsx
Rename renderer/App.tsx to Root.tsx to avoid component name re-use.
