+++ app/index.ts
@@ -1,0 +1,13 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import App from "@foxglove-studio/app/App";
+import ErrorBoundary from "@foxglove-studio/app/components/ErrorBoundary";
+import MultiProvider from "@foxglove-studio/app/components/MultiProvider";
+import { PlayerSourceDefinition } from "@foxglove-studio/app/context/PlayerSelectionContext";
+import ThemeProvider from "@foxglove-studio/app/theme/ThemeProvider";
+
+export { App, ErrorBoundary, MultiProvider, ThemeProvider };
+
+export type { PlayerSourceDefinition };
