+++ package.json
@@ -40,6 +40,7 @@
   "workspaces": {
     "packages": [
       "app",
+      "desktop",
       "packages/**"
     ]
   },
